connecting to cluster state
