+++ src/components/ResourceBrowser/ResourceList/K8SResourceList.tsx
@@ -1,13 +1,9 @@
 import React, { useEffect, useState } from 'react'
 import { useHistory, useLocation, useParams, useRouteMatch } from 'react-router-dom'
-import { ReactComponent as Search } from '../../../assets/icons/ic-search.svg'
-import { ReactComponent as Clear } from '../../../assets/icons/ic-error.svg'
-import { ConditionalWrap, Pagination, Progressing } from '../../common'
+import { Pagination, Progressing } from '../../common'
 import ResourceBrowserActionMenu from './ResourceBrowserActionMenu'
 import {
-    CLUSTER_SELECT_STYLE,
     K8S_RESOURCE_LIST,
-    NAMESPACE_NOT_APPLICABLE_OPTION,
     RESOURCE_EMPTY_PAGE_STATE,
     RESOURCE_LIST_EMPTY_STATE,
     RESOURCE_PAGE_SIZE_OPTIONS,
@@ -15,13 +11,11 @@ import {
 } from '../Constants'
 import { K8SResourceListType } from '../Types'
 import ResourceListEmptyState from './ResourceListEmptyState'
-import ReactSelect from 'react-select'
-import { Option } from '../../../components/v2/common/ReactSelect.utils'
 import AppDetailsStore from '../../v2/appDetails/appDetails.store'
 import { toast } from 'react-toastify'
 import { EventList } from './EventList'
 import Tippy from '@tippyjs/react'
-import { ClusterOptionWithIcon, ResourceValueContainerWithIcon, tippyWrapper } from './ResourceList.component'
+import ResourceFilterOptions from './ResourceFilterOptions'
 
 export function K8SResourceList({
     selectedResource,
@@ -42,10 +36,10 @@ export function K8SResourceList({
     searchApplied,
     setSearchApplied,
     handleFilterChanges,
+    clearSearch,
 }: K8SResourceListType) {
     const { push } = useHistory()
     const { url } = useRouteMatch()
-    const location = useLocation()
     const { clusterId, namespace, nodeType, node } = useParams<{
         clusterId: string
         namespace: string
@@ -77,43 +71,6 @@ export function K8SResourceList({
     const resetPaginator = () => {
         setResourceListOffset(0)
         setPageSize(100)
-    }
-
-    const clearSearch = (): void => {
-        if (searchApplied) {
-            handleFilterChanges('', resourceList)
-            setSearchApplied(false)
-        }
-        setSearchText('')
-    }
-
-    const handleFilterKeyPress = (event): void => {
-        const theKeyCode = event.key
-        if (theKeyCode === 'Backspace' && searchText.length === 1) {
-            clearSearch()
-        } else {
-            handleFilterChanges(event.target.value, resourceList)
-            setSearchApplied(true)
-        }
-    }
-
-    const handleOnChangeSearchText = (event): void => {
-        setSearchText(event.target.value)
-    }
-
-    const handleClusterChange = (selected): void => {
-        onChangeCluster(selected)
-    }
-
-    const handleNamespaceChange = (selected): void => {
-        if (selected.value === selectedNamespace?.value) {
-            return
-        }
-        setSelectedNamespace(selected)
-        handleFilterChanges(searchText, resourceList)
-        push({
-            pathname: location.pathname.replace(`/${namespace}/`, `/${selected.value}/`),
-        })
     }
 
     const handleResourceClick = (e) => {
@@ -148,60 +105,6 @@ export function K8SResourceList({
                 </div>,
             )
         }
-    }
-
-    const renderSearch = (): JSX.Element => {
-        return (
-            <div className="flexbox dc__content-space pt-16 pr-20 pb-12 pl-20">
-                <div className="search dc__position-rel margin-right-0 en-2 bw-1 br-4 h-32">
-                    <Search className="search__icon icon-dim-18" />
-                    <input
-                        type="text"
-                        placeholder={`Search ${selectedResource?.gvk?.Kind || ''}`}
-                        value={searchText}
-                        className="search__input"
-                        onChange={handleOnChangeSearchText}
-                        onKeyUp={handleFilterKeyPress}
-                    />
-                    {searchApplied && (
-                        <button className="search__clear-button" type="button" onClick={clearSearch}>
-                            <Clear className="icon-dim-18 icon-n4 dc__vertical-align-middle" />
-                        </button>
-                    )}
-                </div>
-                <div className="flex">
-                    <ReactSelect
-                        className="w-220"
-                        placeholder="Select Cluster"
-                        options={clusterOptions}
-                        value={selectedCluster}
-                        onChange={handleClusterChange}
-                        styles={CLUSTER_SELECT_STYLE}
-                        components={{
-                            IndicatorSeparator: null,
-                            Option: ClusterOptionWithIcon,
-                            ValueContainer: ResourceValueContainerWithIcon,
-                        }}
-                    />
-                    <ConditionalWrap condition={!selectedResource?.namespaced} wrap={tippyWrapper}>
-                        <ReactSelect
-                            placeholder="Select Namespace"
-                            className="w-220 ml-8"
-                            options={namespaceOptions}
-                            value={selectedResource?.namespaced ? selectedNamespace : NAMESPACE_NOT_APPLICABLE_OPTION}
-                            onChange={handleNamespaceChange}
-                            isDisabled={!selectedResource?.namespaced}
-                            styles={CLUSTER_SELECT_STYLE}
-                            components={{
-                                IndicatorSeparator: null,
-                                Option,
-                                ValueContainer: ResourceValueContainerWithIcon,
-                            }}
-                        />
-                    </ConditionalWrap>
-                </div>
-            </div>
-        )
     }
 
     const renderResourceRow = (resourceData: Record<string, any>, index: number): JSX.Element => {
@@ -361,7 +264,22 @@ export function K8SResourceList({
                 filteredResourceList.length === 0 ? 'no-result-container' : ''
             }`}
         >
-            {renderSearch()}
+            <ResourceFilterOptions
+                selectedResource={selectedResource}
+                clusterOptions={clusterOptions}
+                selectedCluster={selectedCluster}
+                onChangeCluster={onChangeCluster}
+                namespaceOptions={namespaceOptions}
+                selectedNamespace={selectedNamespace}
+                setSelectedNamespace={setSelectedNamespace}
+                searchText={searchText}
+                searchApplied={searchApplied}
+                resourceList={resourceList}
+                setSearchText={setSearchText}
+                setSearchApplied={setSearchApplied}
+                handleFilterChanges={handleFilterChanges}
+                clearSearch={clearSearch}
+            />
             {resourceListLoader ? <Progressing pageLoader /> : renderList()}
         </div>
     )
