+++ src/components/ResourceBrowser/Types.ts
@@ -93,25 +93,30 @@ export interface SidebarType {
     updateResourceSelectionData: (_selected: ApiResourceGroupType) => void
 }
 
-export interface K8SResourceListType {
+export interface ResourceFilterOptionsProps {
     selectedResource: ApiResourceGroupType
     resourceList: ResourceDetailType
-    filteredResourceList: Record<string, any>[]
-    noResults: boolean
     clusterOptions: OptionType[]
     selectedCluster: OptionType
     onChangeCluster: (selectedCluster: OptionType, fromClusterSelect?: boolean) => void
     namespaceOptions: OptionType[]
     selectedNamespace: OptionType
     setSelectedNamespace: React.Dispatch<React.SetStateAction<OptionType>>
-    resourceListLoader: boolean
-    getResourceListData: () => Promise<void>
-    updateNodeSelectionData: (_selected: Record<string, any>) => void
+    hideSearchInput?: boolean
     searchText: string
     setSearchText: React.Dispatch<React.SetStateAction<string>>
     searchApplied: boolean
     setSearchApplied: React.Dispatch<React.SetStateAction<boolean>>
     handleFilterChanges: (_searchText: string, _resourceList: ResourceDetailType) => void
+    clearSearch: () => void
+}
+
+export interface K8SResourceListType extends ResourceFilterOptionsProps {
+    filteredResourceList: Record<string, any>[]
+    noResults: boolean
+    resourceListLoader: boolean
+    getResourceListData: () => Promise<void>
+    updateNodeSelectionData: (_selected: Record<string, any>) => void
 }
 
 export interface ResourceBrowserActionMenuType {
@@ -141,9 +146,8 @@ export interface ClusterOptionType extends OptionType {
     errorInConnecting: string
 }
 
-export interface ConnectingToClusterStateProps {
+export interface ConnectingToClusterStateProps extends ResourceFilterOptionsProps {
     loader: boolean
-    clusterName: string
     errorMsg: string
     handleRetry: (e) => void
     abortController: AbortController
